Move LB related helpers into occm package
